(svn r23201) -Doc: [NoAI] Hide all contents of AIObject from the API documentation as AIs were never able to use it anyway

+++ src/ai/ai_gui.cpp
@@ -294,7 +294,7 @@ struct AISettingsWindow : public Window {
 
 		this->SetWidgetDisabledState(AIS_WIDGET_RESET, _game_mode != GM_MENU && Company::IsValidID(this->slot));
 
-		this->vscroll->SetCount(this->visible_settings.size());
+		this->vscroll->SetCount((int)this->visible_settings.size());
 	}
 
 	/**
